Fix interoperability with responses 0.17.0
