+++ codecov.yml
@@ -1,7 +1,7 @@
 codecov:
   notify:
     # Leave a GitHub comment after all builds have passed
-    after_n_builds: 10
+    after_n_builds: 14
 coverage:
   status:
     project:
