+++ moto/apigateway/models.py
@@ -920,7 +920,7 @@ class RestAPI(CloudFormationModel):
                         content_type="text/plain",
                         match_querystring=False,
                     )
-                    responses_mock._matches.insert(0, callback_response)
+                    responses_mock.add(callback_response)
 
     def create_authorizer(
         self,
